+++ src/common/CutterSeekable.cpp
