+++ src/common/CutterSeekable.h
