+++ src/widgets/DisassemblyWidget.h
@@ -3,7 +3,7 @@
 
 #include "Cutter.h"
 #include "CutterDockWidget.h"
-#include "CutterSeekable.h"
+#include "common/CutterSeekable.h"
 #include "common/RefreshDeferrer.h"
 
 #include <QTextEdit>
